+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [8.22.0] - 2018-10-22
+### Added
+- CXV1-15516 - Adding new SDK Functions for Skills Page (Create, Update Get All, Get).
+- CXV1-15517 - Adding new SDK Functions for Groups Page (Create, Update Get All, Get).
+- CXV1-15503 - Adding new SDK Functions for Users Page (Create, Update Get All, Get).
+- Adding function to get all platform roles predefined.
+- Adding missing functions for Data Access Control.
+
+### Changed
+- Changed function to get a list of historical report folders.
+- Updates at functions for Data Access Control. data-access-group-id parameter added.
+
 ### [8.21.2]
 ### Added
 * CXV1-15326 - Adding code and message from errored API requests as new data to be logged to Kibana